navbar currenuser.name
