+++ src/components/Main/Main.js
@@ -4,7 +4,7 @@ import Preloader from '../Preloader/Preloader';
 import NotFound from '../NotFound/NotFound';
 import { useHistory } from 'react-router-dom';
 
-const Main = ({ loggedin, isSearchHappened, isLoading }) => {
+const Main = ({ loggedin, isSearchHappened, isLoading, handleSigninClick }) => {
   const history = useHistory();
   const savedNewsRoute = history.location.pathname.includes('saved-news');
 
@@ -18,7 +18,11 @@ const Main = ({ loggedin, isSearchHappened, isLoading }) => {
             <h2 className='main__title'>Search Results</h2>
             <ul className='cards__list'>
               <li className='card'>
-                <Card loggedin={loggedin} savedNewsRoute={savedNewsRoute} />
+                <Card
+                  loggedin={loggedin}
+                  savedNewsRoute={savedNewsRoute}
+                  handleSigninClick={handleSigninClick}
+                />
               </li>
               <li className='card'>
                 <Card loggedin={loggedin} savedNewsRoute={savedNewsRoute} />
