--- src/components/utils/auth.js
@@ -1,46 +1,0 @@
-export const BASE_URL = 'http://localhost:3000';
-
-export const register = (email, password, name) => {
-  return fetch(`${BASE_URL}/signup`, {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ email, password, name }),
-  })
-    .then((res) => res.json())
-    .then((res) => res)
-    .catch((err) => console.log(err));
-};
-
-export const login = (email, password) => {
-  return fetch(`${BASE_URL}/signin`, {
-    method: 'POST',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify({ email, password }),
-  })
-    .then((res) => res.json())
-    .then((data) => {
-      localStorage.setItem('jwt', data.token);
-      return checkToken(data.token);
-    })
-    .catch((err) => console.log(err));
-};
-
-export const checkToken = (token) => {
-  return fetch(`${BASE_URL}/users/me`, {
-    method: 'GET',
-    headers: {
-      Accept: 'application/json',
-      'Content-Type': 'application/json',
-      authorization: `Bearer ${token}`,
-    },
-  })
-    .then((res) => res.json())
-    .then((res) => res)
-    .catch((err) => console.log(err));
-};
